+++ src/Fission/Random.hs
@@ -1,3 +1,4 @@
+-- | Random values
 module Fission.Random
   ( byteString
   , text
@@ -11,18 +12,22 @@ import           Data.Word8
 
 import Fission.Internal.Bool (anyX)
 
+-- | Generate random 'Text'
 text :: Natural -> IO Text
 text amount = decodeUtf8Lenient <$> byteString amount
 
+-- | Generate a random 'ByteString'
 byteString :: Natural -> IO ByteString
 byteString amount = BS.filter isURL <$> BS.random amount
 
+-- | Check that a byte represents a valid URL character
 isURL :: Word8 -> Bool
 isURL w = isAsciiUpper w
         || isAsciiLower w
         || isDigit w
         || anyX urlSpecials w
 
+-- | List of URL special character checks
 urlSpecials :: [Word8 -> Bool]
 urlSpecials =
   fmap (==)
